+++ app/renderer/src/main/src/pages/yakitStore/PluginOperator.scss
@@ -8,11 +8,18 @@
     margin-bottom: 0
 }
 
-.edit-plugin-body {
-    position: relative;
+.edit-plugin-title {
+    display: flex;
+    justify-content: space-between;
+    padding: 6px 0 12px;
+    border-bottom: 1px solid #f0f0f0;
+    margin-bottom: 12px;
 
-    .close-icon {
-        position: absolute;
-        right: 0;
+    .title {
+        margin: 0;
+        color: rgba(0, 0, 0, 0.85);
+        font-weight: 500;
+        font-size: 16px;
+        line-height: 22px;
     }
 }
